Create a chart data factory
